moved discussion threads to home page
